week_8/requirements.txt
bcrypt==4.3.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
pgadmin4==9.5
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pytest==8.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
Werkzeug==3.1.3
openai
pgvector
langchain